Advanced pattern matching and removal of old package
